# credit-service/src/main/java/com/example/bank/credits/domain/CreditProductType.java
package com.example.bank.credits.domain;

/**
 * Credit products offered by the bank.
 */
public enum CreditProductType {
    PERSONAL_LOAN,
    BUSINESS_LOAN,
    PERSONAL_CREDIT_CARD,
    BUSINESS_CREDIT_CARD
}